Добавлен интерфейс профиля. Исправлены мелкие недочеты.
