+++ app/src/main/java/com/example/trpp/NavigationActivity/ui/queue/QueueFragment.java
@@ -7,20 +7,27 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.Toast;
+
+
 import com.example.trpp.databinding.FragmentQueueBinding;
 
+
 public class QueueFragment extends Fragment {
+
 
     private FragmentQueueBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+
 
         binding = FragmentQueueBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
         Button to_queue = binding.toQueue;
         Button out_queue = binding.outQueue;
+        EditText input = binding.testFIO;
 
         to_queue.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -38,6 +45,7 @@ public class QueueFragment extends Fragment {
 
         return root;
     }
+
 
     @Override
     public void onDestroyView() {
