+++ app/src/main/res/drawable/oval_button.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="#66DEFC"/>
+            <corners android:radius="60.0dp"/>
+        </shape>
+    </item>
+</selector>
