+++ app/src/main/res/layout/fragment_profile.xml
@@ -1,0 +1,66 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".NavigationActivity.ui.profile.ProfileFragment">
+
+
+    <de.hdodenhof.circleimageview.CircleImageView
+        android:id="@+id/profile_image"
+        android:layout_width="200dp"
+        android:layout_height="200dp"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="30dp"
+        android:src="@drawable/ic_launcher_background"/>
+
+    <EditText
+        android:id="@+id/profile_login"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/profile_image"
+        android:layout_marginHorizontal="50dp"
+        android:layout_marginTop="20dp"
+        android:enabled="false"
+        android:gravity="center"
+        android:text="Some test text"
+        android:textSize="18sp"
+        android:textStyle="bold" />
+
+    <EditText
+        android:id="@+id/profile_FIO"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/profile_login"
+        android:layout_marginHorizontal="50dp"
+        android:layout_marginTop="20dp"
+        android:enabled="false"
+        android:gravity="center"
+        android:text="more some test text"
+        android:textSize="18sp"
+        android:textStyle="bold" />
+
+    <ImageButton
+        android:id="@+id/profile_edit"
+        android:background="@drawable/oval_button"
+        android:layout_width="60dp"
+        android:layout_height="60dp"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentBottom="true"
+        android:layout_marginBottom="80dp"
+        android:src="@drawable/ic_edit_30"
+        android:layout_marginEnd="10dp" />
+
+    <ImageButton
+        android:id="@+id/profile_save"
+        android:background="@drawable/oval_button"
+        android:layout_width="60dp"
+        android:layout_height="60dp"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentBottom="true"
+        android:layout_marginBottom="80dp"
+        android:src="@drawable/ic_save_30"
+        android:layout_marginEnd="10dp"
+        android:visibility="invisible"/>
+</RelativeLayout>
