+++ app/src/main/res/layout/fragment_queue.xml
@@ -25,14 +25,19 @@
         android:layout_marginBottom="80dp"
         android:text="@string/out_queue"/>
 
-    <ScrollView
+    <EditText
+        android:id="@+id/test_FIO"
         android:layout_width="match_parent"
-        android:layout_height="560dp">
+        android:layout_height="wrap_content"
+        android:layout_above="@+id/out_queue"
+        android:layout_marginBottom="20dp"
+        android:hint="Образно я залогинился как" />
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical" />
-    </ScrollView>
+    <ListView
+        android:id="@+id/scroll_view"
+        android:layout_width="match_parent"
+        android:layout_height="520dp">
+
+    </ListView>
 
 </RelativeLayout>
