+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,9 +2,9 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
+        android:id="@+id/navigation_profile"
+        android:icon="@drawable/ic_home_24"
+        android:title="@string/title_profile" />
 
     <item
         android:id="@+id/navigation_dashboard"
